limiters && breeze api

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\PostResource;
 use GuzzleHttp\Psr7\Message;
 use Illuminate\Http\Request;
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -16,7 +17,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate(2));
+        $user=request()->user();
+        $posts=$user->posts()->get();
+        return PostResource::collection($posts);
         }
 
     /**
@@ -25,7 +28,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
    $data= $request->validated();
-   $data['author_id'] = 1; // Assuming the author_id is 1 for demonstration purposes
+   $data['author_id'] = $request->user()->id; // Assuming the author_id is the authenticated user's ID
    $post= Post::create($data);
    return new PostResource($post);
     
@@ -46,6 +49,10 @@ class PostController extends Controller
     public function show(Post $post)//id or use route model binding
     {
        // $post = Post::findOrFail($id);
+       $user = request()->user(); 
+       //if ($user->id !== $post->author_id) {
+        abort_if(Auth::id() != $post->author_id, 403, 'Unauthorized access to this post');
+     //  
         return 
         response()->json([
             "message" => "Post retrieved successfully",
@@ -60,6 +67,8 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+     abort_if(Auth::id() != $post->author_id, 403, 'Unauthorized access to this post');
+
         $data = $request->validated();
         $post->update($data);
         return response()->json([
@@ -78,6 +87,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Unauthorized access to this post');
         $post->delete();
         //return response()->noContent();
     }
